backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-multipart==0.0.22
minio==7.2.20
python-dotenv==1.0.1
email-validator==2.2.0
